move the background mode to a service
