+++ app/src/main/AndroidManifest.xml
@@ -283,6 +283,10 @@
             android:name=".services.UpdateService"
             android:enabled="true"
             android:exported="false" />
+
+        <service android:name=".services.BackgroundMode"
+            android:enabled="true"
+            android:exported="false" />
     </application>
 
 </manifest>
