+++ app/src/main/java/com/github/libretube/fragments/PlayerFragment.kt
@@ -58,7 +58,7 @@ import com.github.libretube.obj.Streams
 import com.github.libretube.obj.Subscribe
 import com.github.libretube.preferences.PreferenceHelper
 import com.github.libretube.preferences.PreferenceKeys
-import com.github.libretube.util.BackgroundMode
+import com.github.libretube.services.BackgroundMode
 import com.github.libretube.util.ConnectionHelper
 import com.github.libretube.util.CronetHelper
 import com.github.libretube.util.DescriptionAdapter
@@ -528,12 +528,13 @@ class PlayerFragment : Fragment() {
             exoPlayer.pause()
 
             // start the background mode
-            BackgroundMode
-                .getInstance()
-                .playOnBackgroundMode(
-                    requireContext(),
-                    videoId!!
-                )
+            if (Globals.backgroundModeIntent != null) {
+                activity?.stopService(Globals.backgroundModeIntent)
+            }
+            val intent = Intent(context, BackgroundMode::class.java)
+            intent.putExtra("videoId", videoId)
+            Globals.backgroundModeIntent = intent
+            activity?.startService(intent)
         }
 
         binding.playerScrollView.viewTreeObserver
